fix : userId plannerId Long으로 변경

+++ weddingyou_backend/src/main/java/com/mysite/weddingyou_backend/planner/Planner.java
@@ -21,7 +21,7 @@ public class Planner {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "planner_id")
-    private int plannerId;
+    private Long plannerId;
 
     @Column(name = "name", nullable = false)
     private String name;
