+++ weddingyou_backend/src/main/java/com/mysite/weddingyou_backend/planner/PlannerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlannerRepository extends JpaRepository<Planner, Integer> {
+public interface PlannerRepository extends JpaRepository<Planner, Long> {
     Planner findByEmail(String email);
 }
